LayoutTests/platform/ios/http/tests/cookies/same-site/popup-same-site-via-cross-site-redirect-expected.txt


FAIL '127.0.0.1' is same-site with itself, so samesite cookies are sent. assert_equals: implicit-strict expected (undefined) undefined but got (string) "1"
